I renamed a variable in the shooter file
Before we run this, we need to figure out what joystick this is mapped to. Until then, I will try to figure out how to map this to a trigger/button. Before deploying to the robot, delete either the drive program or the shooter program since controls will likely interfere.